Refactoring: moved stack license sanity checks into new module
